backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-magic==0.4.27
pytesseract==0.3.13
Pillow==11.2.1
google-generativeai==0.8.5
pydantic==2.11.3
SQLAlchemy==2.0.24
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-dotenv==1.0.0
psycopg2-binary
python-dotenv